LPS-55928 move suggest related classes into proper package
