+++ modules/portal/portal-search-lucene/src/com/liferay/portal/search/lucene/internal/LuceneSpellCheckIndexWriter.java
@@ -16,15 +16,15 @@ package com.liferay.portal.search.lucene.internal;
 
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.search.BaseSpellCheckIndexWriter;
 import com.liferay.portal.kernel.search.DictionaryEntry;
 import com.liferay.portal.kernel.search.DictionaryReader;
 import com.liferay.portal.kernel.search.DocumentImpl;
-import com.liferay.portal.kernel.search.NGramHolder;
-import com.liferay.portal.kernel.search.NGramHolderBuilderUtil;
 import com.liferay.portal.kernel.search.SearchContext;
 import com.liferay.portal.kernel.search.SearchException;
-import com.liferay.portal.kernel.search.SuggestionConstants;
+import com.liferay.portal.kernel.search.suggest.BaseSpellCheckIndexWriter;
+import com.liferay.portal.kernel.search.suggest.NGramHolder;
+import com.liferay.portal.kernel.search.suggest.NGramHolderBuilderUtil;
+import com.liferay.portal.kernel.search.suggest.SuggestionConstants;
 import com.liferay.portal.kernel.util.StringPool;
 
 import java.io.IOException;
