+++ portal-impl/src/META-INF/search-spring.xml
@@ -38,7 +38,7 @@
 	<bean id="com.liferay.portal.kernel.repository.search.RepositorySearchQueryBuilderUtil" class="com.liferay.portal.kernel.repository.search.RepositorySearchQueryBuilderUtil">
 		<property name="repositorySearchQueryBuilder" ref="com.liferay.portal.kernel.repository.search.RepositorySearchQueryBuilder" />
 	</bean>
-	<bean id="com.liferay.portal.kernel.search.CollatorUtil" class="com.liferay.portal.kernel.search.CollatorUtil">
+	<bean id="com.liferay.portal.kernel.search.suggest.CollatorUtil" class="com.liferay.portal.kernel.search.suggest.CollatorUtil">
 		<property name="collator">
 			<bean class="com.liferay.portal.search.DefaultCollatorImpl" />
 		</property>
