+++ portal-service/src/com/liferay/portal/kernel/search/BaseIndexWriter.java
@@ -16,6 +16,7 @@ package com.liferay.portal.kernel.search;
 
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.kernel.search.suggest.SpellCheckIndexWriter;
 
 /**
  * @author Michael C. Han
