+++ portal-service/src/com/liferay/portal/kernel/search/IndexSearcher.java
@@ -16,6 +16,7 @@ package com.liferay.portal.kernel.search;
 
 import com.liferay.portal.kernel.messaging.proxy.MessagingProxy;
 import com.liferay.portal.kernel.messaging.proxy.ProxyMode;
+import com.liferay.portal.kernel.search.suggest.QuerySuggester;
 
 /**
  * @author Bruno Farache
