+++ portal-service/src/com/liferay/portal/kernel/search/QueryIndexingHitsProcessor.java
@@ -14,6 +14,8 @@
 
 package com.liferay.portal.kernel.search;
 
+import com.liferay.portal.kernel.search.suggest.SuggestionConstants;
+
 import java.util.Locale;
 
 /**
