+++ portal-service/src/com/liferay/portal/kernel/search/suggest/BaseQuerySuggester.java
@@ -12,7 +12,13 @@
  * details.
  */
 
-package com.liferay.portal.kernel.search;
+package com.liferay.portal.kernel.search.suggest;
+
+import com.liferay.portal.kernel.search.DocumentImpl;
+import com.liferay.portal.kernel.search.Field;
+import com.liferay.portal.kernel.search.SearchContext;
+import com.liferay.portal.kernel.search.SearchException;
+import com.liferay.portal.kernel.search.TokenizerUtil;
 
 import java.util.List;
 import java.util.Map;
