+++ portal-service/src/com/liferay/portal/kernel/search/suggest/KeywordsSuggestionHolder.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.kernel.search;
+package com.liferay.portal.kernel.search.suggest;
 
 import com.liferay.portal.kernel.util.Validator;
