+++ portal-service/src/com/liferay/portal/kernel/search/suggest/NGramHolderBuilderUtil.java
@@ -12,8 +12,9 @@
  * details.
  */
 
-package com.liferay.portal.kernel.search;
+package com.liferay.portal.kernel.search.suggest;
 
+import com.liferay.portal.kernel.search.SearchException;
 import com.liferay.registry.Registry;
 import com.liferay.registry.RegistryUtil;
 import com.liferay.registry.ServiceTracker;
