+++ portal-service/src/com/liferay/portal/kernel/search/suggest/SpellCheckIndexWriter.java
@@ -12,7 +12,10 @@
  * details.
  */
 
-package com.liferay.portal.kernel.search;
+package com.liferay.portal.kernel.search.suggest;
+
+import com.liferay.portal.kernel.search.SearchContext;
+import com.liferay.portal.kernel.search.SearchException;
 
 /**
  * @author Michael C. Han
