+++ portal-service/src/com/liferay/portal/kernel/search/suggest/SuggestionConstants.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.kernel.search;
+package com.liferay.portal.kernel.search.suggest;
 
 /**
  * @author Michael C. Han
